UPDATE: all contract and propay realted entities got necessary attributes

+++ src/main/java/de/hhu/propra/sharingplatform/model/Item.java
@@ -1,18 +1,12 @@
 package de.hhu.propra.sharingplatform.model;
 
 
-import java.util.ArrayList;
-import java.util.List;
-import javax.persistence.CascadeType;
-import javax.persistence.Entity;
-import javax.persistence.GeneratedValue;
-import javax.persistence.GenerationType;
-import javax.persistence.Id;
-import javax.persistence.ManyToOne;
-import javax.persistence.OneToMany;
-
 import lombok.Data;
 import lombok.ToString;
+
+import javax.persistence.*;
+import java.util.ArrayList;
+import java.util.List;
 
 @Data
 @Entity
@@ -25,8 +19,8 @@ public class Item {
 
     private String name;
     private String description;
-    private Integer bail;
-    private Integer price; // each day
+    private Double bail;
+    private Double price; // each day
     private boolean available;
     private String location; // maybe change to java location class
     private boolean deleted;
